Added main.py
This is a non-functional, work-in-progress update to the repository. This commit adds the main code so that others may see how the root board generation is going to function, and see the beginning stages of the GUI development. That being said, it is nowhere near complete, and cannot be used to actually generate any boards yet.