Updated main.py and added developmentNotes.txt
This update contains a major overhaul to the GUI, completely redesigning it and using a matching, legible color pallet. It may not seem like it from the surface, but the changes are huge. The design scheme has been written in such a way as to allow custom color templates in the future, and much more. Additionally, support for custom character sets has been added, and many core runtime errors within the board generation process have been eliminated. There is still a lot of work to go before the code is perfected, or even usable, but the end is now in site. One final note: a separate file has been added to contain notes and future ideas.