Updated main.py
The board generation process is finally functional, and can now be used to successfully generate game boards with the provided vocabulary! There is definitely some unnecessary code within the genBoards function, but that can be cleaned-up later. For now, I'm glad that I was finally able to get it to work. As it is, the boards will currently only output to the commandline, therefore the application is not yet ready for standard use or distribution. That being said, it's very close to being ready. All that is left to make it functional is to add the PDF generation, and to fill the default text with random characters from the chosen charset.